+++ libs/dot-components/src/lib/components/radio/RadioButton.styles.ts
@@ -1,0 +1,12 @@
+import { Radio } from '@material-ui/core';
+import styled, { css } from 'styled-components';
+
+export const rootClassName = 'dot-radio';
+
+export const StyledRadioButton = styled(Radio)`
+  ${({ theme }) => css`{
+  &.${rootClassName} {
+    padding: 8px;
+  }
+  `}
+`;