+++ CHANGE_LOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## Current
+
+- **RadionButton**: added NEW `RadioButton` component
+- **RadioGroup**: added NEW `RadioGroup` component
+- **RadionButton**: moved to Components
+- **RadioGroup**: moved to Components
+
 ## 0.0.0-alpha.8
 
 - **AppToolbar**: fixed postion styles so the AppTollbar is always flush to top, left, and right